Update wallet dependencey to v4.0.27

+++ lib-impl/src/main/java/org/ergoplatform/appkit/impl/ErgoProverBuilderImpl.scala
@@ -77,7 +77,7 @@ class ErgoProverBuilderImpl(_ctx: BlockchainContextBase) extends ErgoProverBuild
       override def inputCost: Int = _params.getInputCost
       override def dataInputCost: Int = _params.getDataInputCost
       override def outputCost: Int = _params.getOutputCost
-      override def maxBlockCost: Long = _params.getMaxBlockCost
+      override def maxBlockCost: Int = _params.getMaxBlockCost
       override def softForkStartingHeight: Option[Int] = ???
       override def softForkVotesCollected: Option[Int] = ???
       override def blockVersion: Byte = _params.getBlockVersion.byteValue
